requirements.txt
streamlit
azure-storage-blob
dotenv
pyodbc